pyproject.toml
[tool.poetry]
name = "tanos"
version = "0.1.0"
description = ""
authors = ["noah <noah@apartalk.com>"]

[tool.poetry.dependencies]
python = "^3.8"
flask = "*"
flask-sqlalchemy = "*"
flask-migrate = "*"
flask-jwt-extended = "3.25.0"
psycopg2-binary = "*"
importlib-metadata = "*"
inject = "*"
pydantic = "*"
pypubsub = "*"
flasgger = "*"
boto3= "1.12.17"
gevent = "^20.1.0"
gunicorn = "==20.0.4"
requests = "^2.25.1"
redis = "^3.5.3"
newrelic = "*"
sentry-sdk = "^1.1.0"
GeoAlchemy2 = "^0.9.1"
geojson-pydantic = "^0.3.0"
supervisor = "^4.1.0"

[tool.poetry.dev-dependencies]
black = "^19.10b0"
pytest-xdist = "*"
pytest = "*"
pytest-factoryboy = "*"
faker = "*"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
